compute pressure: Process sampleRate option in Blink
This patch implements the processing of the PressureObserverOptions
`sampleRate` by checking it validity.

The user notification of a PressureRecord change based on the
comparison between the time difference between 2 platform collector
updates and the requested sample period is also implemented in this
patch. The user will be only updated at the observer requested sample
rate, even if platform collector updates are more frequent.

Bug: 1371073
Change-Id: I1782334926320dd1079f9ad1e6e55a89f538e0ec