[FlexNG] OOF static position when CB is outside multicol
The static position of OOFs in a fragmented flex container are
determined within the last fragment. If the OOF's CB lives outside the
associated multicol, then that static position will be relative to the
last parent fragment and would become negative. Adjust the static
position once it is propagated to the multicol to account for the
potential negative offset (only if the OOF hasn't found its CB yet).

Note that external\wpt\css\css-break\flexbox\
flex-container-fragmentation-011.html is not working as expected. The
main reason being that the offset will get adjusted once we know the
OOF's block size in this case, which happens after we've adjusted to
account for any negative block offset.

This solution likely won't work if there is nested fragmentation
involved, such that the multicol/flex container involved are both
fragmented.

This problem applies to grid, as well, but this will be handled in a
separate CL.

Bug: 1367912
Change-Id: I0f0171d50aee4e0c2f582047620a3ac9e7b35fd3