[bfcache] Do not reset notRestoredReasons upon successful bfcache navigation
This CL makes PerformanceNavigationEntry.notRestoredReasons not reset
after successful bfcache navigation. After this CL, when the page is
successfully restored from bfcache, the reasons from the previous
navigation stay.

Now, the only time we reset the reasons is when we get a new set of
reasons.

This was discussed on the email thread here:
https://groups.google.com/a/chromium.org/g/bfcache-dev/c/fGPWPVTbyJA

Bug:1363438

Change-Id: I5214f56e8424356cad8190c2a0144e4d1c017746